[SC-71] refactor : common directory로 Department 이동
