+++ src/main/java/inu/codin/codin/domain/info/office/service/OfficeService.java
@@ -4,7 +4,7 @@ import inu.codin.codin.domain.info.office.dto.OfficeListResDTO;
 import inu.codin.codin.domain.info.office.dto.OfficeMemberResDTO;
 import inu.codin.codin.domain.info.office.entity.Office;
 import inu.codin.codin.domain.info.office.repository.OfficeRepository;
-import inu.codin.codin.domain.user.entity.Department;
+import inu.codin.codin.common.Department;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
