+++ src/main/java/inu/codin/codin/domain/info/professor/controller/ProfessorController.java
@@ -3,7 +3,7 @@ package inu.codin.codin.domain.info.professor.controller;
 import inu.codin.codin.domain.info.professor.service.ProfessorService;
 import inu.codin.codin.domain.info.professor.dto.ProfessorListResDTO;
 import inu.codin.codin.domain.info.professor.dto.ProfessorThumbnailResDTO;
-import inu.codin.codin.domain.user.entity.Department;
+import inu.codin.codin.common.Department;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
