+++ src/main/java/inu/codin/codin/domain/info/professor/dto/ProfessorListResDTO.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.info.professor.dto;
 
 import inu.codin.codin.domain.info.professor.entity.Professor;
-import inu.codin.codin.domain.user.entity.Department;
+import inu.codin.codin.common.Department;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
