+++ src/main/java/inu/codin/codin/domain/info/professor/service/ProfessorService.java
@@ -5,7 +5,7 @@ import inu.codin.codin.domain.info.professor.dto.ProfessorThumbnailResDTO;
 import inu.codin.codin.domain.info.professor.entity.Professor;
 import inu.codin.codin.domain.info.professor.exception.ProfessorNotFoundException;
 import inu.codin.codin.domain.info.professor.repository.ProfessorRepository;
-import inu.codin.codin.domain.user.entity.Department;
+import inu.codin.codin.common.Department;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
